+++ packages/solutions/app-tools/lib/types.d.ts
@@ -53,8 +53,12 @@ declare module '*.svg' {
     React.SVGProps<SVGSVGElement>
   >;
 
-  const src: string;
-  export default src;
+  /**
+   * The default export type depends on the svgDefaultExport config,
+   * it can be a string or a ReactComponent
+   * */
+  const content: any;
+  export default content;
 }
 
 declare module '*.bmp?inline' {
@@ -99,12 +103,8 @@ declare module '*.svg?inline' {
     React.SVGProps<SVGSVGElement>
   >;
 
-  /**
-   * The default export type depends on the svgDefaultExport config,
-   * it can be a string or a ReactComponent
-   * */
-  const content: any;
-  export default content;
+  const src: string;
+  export default src;
 }
 
 declare module '*.bmp?url' {