+++ tests/e2e/builder/cases/inline-chunk/index.test.ts
@@ -1,8 +1,8 @@
 import path from 'path';
 import { expect } from '@modern-js/e2e/playwright';
-import { RUNTIME_CHUNK_NAME } from '@modern-js/builder-shared';
 import { webpackOnlyTest } from '@scripts/helper';
-import { build } from '@scripts/shared';
+import { build, getHrefByEntryName } from '@scripts/shared';
+import { RUNTIME_CHUNK_NAME } from '@modern-js/builder-shared';
 
 // todo: Rspack not output RUNTIME_CHUNK_NAME.js.map
 const isRuntimeChunkInHtml = (html: string): boolean =>
@@ -20,7 +20,7 @@ const toolsConfig = {
   },
 };
 
-webpackOnlyTest('inline runtime chunk by default', async () => {
+webpackOnlyTest('inline runtime chunk by default', async ({ page }) => {
   const builder = await build(
     {
       cwd: __dirname,
@@ -29,8 +29,13 @@ webpackOnlyTest('inline runtime chunk by default', async () => {
     {
       tools: toolsConfig,
     },
-    false,
   );
+
+  // test runtime
+  await page.goto(getHrefByEntryName('index', builder.port));
+
+  expect(await page.evaluate(`window.test`)).toBe('aaaa');
+
   const files = await builder.unwrapOutputJSON(false);
 
   // no builder-runtime file in output
@@ -93,37 +98,46 @@ webpackOnlyTest(
   },
 );
 
-webpackOnlyTest('inline all scripts and emit all source maps', async () => {
-  const builder = await build(
-    {
-      cwd: __dirname,
-      entry: {
-        index: path.resolve(__dirname, './src/index.js'),
-        another: path.resolve(__dirname, './src/another.js'),
+webpackOnlyTest(
+  'inline all scripts and emit all source maps',
+  async ({ page }) => {
+    const builder = await build(
+      {
+        cwd: __dirname,
+        entry: {
+          index: path.resolve(__dirname, './src/index.js'),
+          another: path.resolve(__dirname, './src/another.js'),
+        },
       },
-    },
-    {
-      output: {
-        enableInlineScripts: true,
+      {
+        output: {
+          enableInlineScripts: true,
+        },
+        tools: toolsConfig,
       },
-      tools: toolsConfig,
-    },
-    false,
-  );
-  const files = await builder.unwrapOutputJSON(false);
+    );
 
-  // no entry chunks or runtime chunks in output
-  expect(
-    Object.keys(files).filter(
-      fileName => fileName.endsWith('.js') && !fileName.includes('/async/'),
-    ).length,
-  ).toEqual(0);
+    await page.goto(getHrefByEntryName('index', builder.port));
 
-  // all source maps in output
-  expect(
-    Object.keys(files).filter(fileName => fileName.endsWith('.js.map')).length,
-  ).toEqual(5);
-});
+    // test runtime
+    expect(await page.evaluate(`window.test`)).toBe('aaaa');
+
+    const files = await builder.unwrapOutputJSON(false);
+
+    // no entry chunks or runtime chunks in output
+    expect(
+      Object.keys(files).filter(
+        fileName => fileName.endsWith('.js') && !fileName.includes('/async/'),
+      ).length,
+    ).toEqual(0);
+
+    // all source maps in output
+    expect(
+      Object.keys(files).filter(fileName => fileName.endsWith('.js.map'))
+        .length,
+    ).toEqual(5);
+  },
+);
 
 webpackOnlyTest('using RegExp to inline scripts', async () => {
   const builder = await build(