+++ tests/integration/css/fixtures/single-less/src/App.jsx
@@ -1,3 +1,0 @@
-import './app.less';
-
-export default () => 'single-less';