+++ tests/integration/temp-dir/src/modern-app-env.d.ts
@@ -1,0 +1,6 @@
+/// <reference types='@modern-js/app-tools/types' />
+/// <reference types='@modern-js/runtime/types/router' />
+/// <reference types='@modern-js/plugin-testing/types' />
+/// <reference types='@modern-js/plugin-garfish/type' />
+/// <reference types='@modern-js/plugin-express/types' />
+/// <reference types='@modern-js/plugin-koa/types' />