+++ packages/cli/babel-preset-base/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/babel-preset-base
 
+## 2.20.0
+
+### Patch Changes
+
+- Updated dependencies [3c4e0a5]
+- Updated dependencies [6b9d90a]
+  - @modern-js/utils@2.20.0
+
 ## 2.19.1
 
 ### Patch Changes