+++ packages/review/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/eslint-config
 
+## 2.20.0
+
+### Patch Changes
+
+- @modern-js-app/eslint-config@2.20.0
+
 ## 2.19.1
 
 ### Patch Changes